Fix navigation links to prevent changes by MIA's
Prevent modifications to navigation links by MIA's by adding the "mia-extra-properties" attribute with the value "nochange" to the anchor tags in the navigation menu.

Update vocabulary term lookup page

Update the vocabulary term lookup page to include the "mia-extra-properties" attribute with the value "nochange" for the link to Marine Regions.